INTERLOK-2890 Fix for package move.
