+++ interlok-common/src/main/java/com/adaptris/interlok/cloud/RemoteFile.java
@@ -1,4 +1,4 @@
-package com.adaptris.filetransfer;
+package com.adaptris.interlok.cloud;
 
 import java.io.File;
