+++ interlok-common/src/test/java/com/adaptris/interlok/cloud/RemoteFileTest.java
@@ -1,4 +1,4 @@
-package com.adaptris.filetransfer;
+package com.adaptris.interlok.cloud;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
