+++ interlok-core/src/main/java/com/adaptris/sftp/SftpClient.java
@@ -33,8 +33,8 @@ import org.apache.commons.lang3.StringUtils;
 import com.adaptris.filetransfer.FileTransferClient;
 import com.adaptris.filetransfer.FileTransferClientImp;
 import com.adaptris.filetransfer.FileTransferException;
-import com.adaptris.filetransfer.RemoteFile;
 import com.adaptris.ftp.FtpException;
+import com.adaptris.interlok.cloud.RemoteFile;
 import com.adaptris.util.FifoMutexLock;
 import com.jcraft.jsch.Channel;
 import com.jcraft.jsch.ChannelSftp;
